runtime permission

+++ app/src/main/java/com/ro0kiey/igank/ui/activity/MeiziActivity.java
@@ -1,10 +1,17 @@
 package com.ro0kiey.igank.ui.activity;
 
+import android.Manifest;
+import android.app.Activity;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.Environment;
+import android.support.annotation.NonNull;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.ContextCompat;
+import android.support.v4.content.PermissionChecker;
 import android.support.v4.view.ViewCompat;
 import android.util.Log;
 import android.view.Menu;
@@ -27,7 +34,6 @@ public class MeiziActivity extends BaseActivity {
 
     private ImageView imageView;
     private String meiziUrl;
-    private boolean hasSaved = false;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -64,13 +70,7 @@ public class MeiziActivity extends BaseActivity {
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()){
             case R.id.save_meizi:
-                if (!hasSaved){
-                    saveMeiziImage(meiziUrl);
-                    ToastUtils.SnackBarShort(imageView, R.string.save_success);
-                    hasSaved = true;
-                } else {
-                    ToastUtils.SnackBarShort(imageView, R.string.already_saved);
-                }
+                checkPermission();
                 break;
             default:
                 break;
@@ -78,24 +78,61 @@ public class MeiziActivity extends BaseActivity {
         return false;
     }
 
+    private void checkPermission(){
+        if(ContextCompat.checkSelfPermission(MeiziActivity.this, Manifest.permission.READ_EXTERNAL_STORAGE)
+                != PackageManager.PERMISSION_GRANTED){
+            ActivityCompat.requestPermissions(MeiziActivity.this, new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE}, 1);
+        } else {
+            saveMeiziImage(meiziUrl);
+        }
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        switch (requestCode){
+            case 1:
+                if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED){
+                    saveMeiziImage(meiziUrl);
+                } 
+                break;
+            default:
+                break;
+        }
+    }
+
     private void saveMeiziImage(String url) {
-        Bitmap bitmap = imageView.getDrawingCache();
-        File appDir = new File(Environment.getExternalStorageDirectory(), "igank");
-        if (!appDir.exists()){
-            appDir.mkdirs();
+        if (ExistSDCard()){
+            Bitmap bitmap = imageView.getDrawingCache();
+            File appDir = new File(Environment.getExternalStorageDirectory().getPath(), "igank/");
+            if (!appDir.exists()){
+                appDir.mkdirs();
+            }
+            String fileName = url.substring(8, url.length() - 4).replace("/", "-") + ".jpg";
+            File file = new File(appDir, fileName);
+            if (!file.exists()){
+                try {
+                    FileOutputStream os = new FileOutputStream(file);
+                    bitmap.compress(Bitmap.CompressFormat.JPEG, 100, os);
+                    os.flush();
+                    os.close();
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+                this.sendBroadcast(new Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, Uri.fromFile(file)));
+                ToastUtils.SnackBarShort(imageView, R.string.save_success);
+            } else {
+                ToastUtils.SnackBarShort(imageView, R.string.already_saved);
+            }
+            //imageView.setDrawingCacheEnabled(false);
+            Log.d("MeiziActivity", "saveMeiziImage: ");
         }
-        String fileName = url.substring(8).replace("/", "-") + "girl.jpg";
-        File file = new File(appDir, fileName);
-        try {
-            FileOutputStream os = new FileOutputStream(file);
-            bitmap.compress(Bitmap.CompressFormat.JPEG, 100, os);
-            os.flush();
-            os.close();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        //imageView.setDrawingCacheEnabled(false);
-        Log.d("MeiziActivity", "saveMeiziImage: ");
-        this.sendBroadcast(new Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, Uri.fromFile(file)));
+    }
+
+    private boolean ExistSDCard() {
+        if (android.os.Environment.getExternalStorageState().equals(
+                android.os.Environment.MEDIA_MOUNTED)) {
+            return true;
+        } else
+            return false;
     }
 }
